requirements.txt
Flask==2.3.3
tensorflow==2.15.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
pandas==2.0.3
scikit-image==0.21.0
segmentation-models-pytorch==0.3.3
scikit-learn==1.3.0
Werkzeug==2.3.7
requests==2.31.0
